Pull out video visualization component
